Fix listeners clearing at page creations. Navigation to global room from pageBase class, + fix some errors

+++ src/stream/stream.js
@@ -231,8 +231,8 @@ class Stream {
 
 	stop() {
 		this.isPlalistInitialized = false;
-		this.ffmpegProc.kill()
-
+		if(this.ffmpegProc)
+			this.ffmpegProc.kill()
 		if( this.watcherPID )
 			this.watcherPID.close()
 		this.blocks = [];
